Revert "overlay: increase mms size limit"
This value is rejected by multiple carriers. 300K is a safe limit, 1MB isn't, 1.2MB even less. 2 out of 3 carriers in my region reject messages due to this.

Please re-submit this overlayed for mcc/mnc pairs of carriers in which this is _known_ to be working.

An attempt to do this at a device level (and following revert) was attempted in the past, see here: http://review.cyanogenmod.org/#/c/73987/

This reverts commit 2df1e3dd2572eb485bb9ef3af97325ea80974349.

Change-Id: Ie0558f8c896c2fde1c7e956c9af79198b208b23e